Commit old sites 2005-2016
These are the SPA sites from previous years (accessible via e.g. http://www.spaconference.org/spa2005/). These had been part of the spa conference repo, but in fact there is no need for them to be as once the year has changed the old site is not touched, it is just there so that it can still be viewed. So it doesn't  make sense to have them all as part of the main repo.