Commit files for basic 2022 site
This is not actually a previous site, it's the current one, but the idea is that we link out to a miniSPA site. I could just replace everthing in the deprecated `spa-website` repo with this, but not clear what the next steps are for SPA so I think this makes more sense for now.